Configurado detales de acesso à criação de oferta de monitoria ao professor e estudante (tanto logado quanto usuário externo

+++ db/migrate/20161210121327_create_disciplinas.rb
@@ -4,7 +4,7 @@ class CreateDisciplinas < ActiveRecord::Migration[5.0]
       t.string :nome
       t.string :codigo
       t.string :professor
-      t.decimal :vagas
+      t.integer :vagas
       t.timestamps
     end
   end
